+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Update Kaniko container builder to 1.11.0
 * Add support for _Kafka node pools_ according to [Strimzi Proposal #50](https://github.com/strimzi/proposals/blob/main/050-Kafka-Node-Pools.md)
 * Update OpenTelemetry 1.19.0
+* Fixed ordering of JVM performance options [#8579](https://github.com/strimzi/strimzi-kafka-operator/issues/8579)
 
 ### Changes, deprecations and removals